requirements.txt
lxml>=4.9.1
beautifulsoup4>=4.11.1
typing-extensions>=4.4.0
tqdm>=4.64.1
retrying>=1.3.4
pillow>=9.4.0
opencv-python>=4.7.0.72
opencv-contrib-python>=4.7.0.72
transformers>=4.26.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
python-dotenv>=1.0.0
pandas>=1.5.3
scikit-learn>=1.2.1
matplotlib>=3.6.3
seaborn>=0.12.2
wandb>=0.13.10
#flash-attn>=0.2.8